requirements.txt
certifi==2016.2.28
coverage==4.4.1
Django==1.11.5
django-debug-toolbar==1.8
py==1.4.34
pytest==3.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
